+++ opennebula/resource_opennebula_virtual_machine.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"reflect"
 	"strconv"
 	"strings"
 	"time"
@@ -15,7 +14,6 @@ import (
 	"github.com/OpenNebula/one/src/oca/go/src/goca"
 	dyn "github.com/OpenNebula/one/src/oca/go/src/goca/dynamic"
 	errs "github.com/OpenNebula/one/src/oca/go/src/goca/errors"
-	"github.com/OpenNebula/one/src/oca/go/src/goca/schemas/shared"
 	"github.com/OpenNebula/one/src/oca/go/src/goca/schemas/vm"
 	vmk "github.com/OpenNebula/one/src/oca/go/src/goca/schemas/vm/keys"
 )
@@ -113,159 +111,15 @@ func resourceOpennebulaVirtualMachine() *schema.Resource {
 				Computed:    true,
 				Description: "Current LCM state of the VM",
 			},
-			"cpu": {
-				Type:        schema.TypeFloat,
-				Optional:    true,
-				Computed:    true,
-				Description: "Amount of CPU quota assigned to the virtual machine",
-			},
-			"vcpu": {
-				Type:        schema.TypeInt,
-				Optional:    true,
-				Computed:    true,
-				Description: "Number of virtual CPUs assigned to the virtual machine",
-			},
-			"memory": {
-				Type:        schema.TypeInt,
-				Optional:    true,
-				Computed:    true,
-				Description: "Amount of memory (RAM) in MB assigned to the virtual machine",
-			},
-			"context": {
-				Type:        schema.TypeMap,
-				Optional:    true,
-				Computed:    true,
-				Description: "Context variables",
-			},
-			"disk": {
-				Type:        schema.TypeList,
-				Optional:    true,
-				Computed:    true,
-				Description: "Definition of disks assigned to the Virtual Machine",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"image_id": {
-							Type:     schema.TypeInt,
-							Required: true,
-						},
-						"size": {
-							Type:     schema.TypeInt,
-							Computed: true,
-							Optional: true,
-						},
-						"target": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"driver": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-					},
-				},
-			},
-			"graphics": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Computed:    true,
-				MinItems:    0,
-				MaxItems:    1,
-				Description: "Definition of graphics adapter assigned to the Virtual Machine",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"listen": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"port": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"type": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"keymap": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Default:  "en-us",
-						},
-					},
-				},
-			},
-			"nic": {
-				Type:        schema.TypeList,
-				Optional:    true,
-				Computed:    true,
-				Description: "Definition of network adapter(s) assigned to the Virtual Machine",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"ip": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"mac": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"model": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"network_id": {
-							Type:     schema.TypeInt,
-							Required: true,
-						},
-						"network": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"physical_device": {
-							Type:     schema.TypeString,
-							Computed: true,
-							Optional: true,
-						},
-						"security_groups": {
-							Type:     schema.TypeList,
-							Optional: true,
-							Computed: true,
-							Elem: &schema.Schema{
-								Type: schema.TypeInt,
-							},
-						},
-						"nic_id": {
-							Type:     schema.TypeInt,
-							Computed: true,
-						},
-					},
-				},
-			},
-			"os": {
-				Type:     schema.TypeSet,
-				Optional: true,
-				Computed: true,
-				MaxItems: 1,
-				//Computed:    true,
-				Description: "Definition of OS boot and type for the Virtual Machine",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"arch": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"boot": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-					},
-				},
-			},
+			"cpu":      cpuSchema(),
+			"vcpu":     vcpuSchema(),
+			"memory":   memorySchema(),
+			"context":  contextSchema(),
+			"disk":     diskSchema(),
+			"graphics": graphicsSchema(),
+			"nic":      nicSchema(),
+			"os":       osSchema(),
+			"vmgroup":  vmGroupSchema(),
 			"ip": {
 				Type:        schema.TypeString,
 				Computed:    true,
@@ -275,27 +129,6 @@ func resourceOpennebulaVirtualMachine() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 				Description: "Name of the Group that onws the VM, If empty, it uses caller group",
-			},
-			"vmgroup": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Computed:    true,
-				ForceNew:    true,
-				MinItems:    0,
-				MaxItems:    1,
-				Description: "Virtual Machine Group to associate with during VM creation only. If it changes, a New VM is created",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"vmgroup_id": {
-							Type:     schema.TypeInt,
-							Required: true,
-						},
-						"role": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-					},
-				},
 			},
 		},
 	}
@@ -479,137 +312,6 @@ func resourceOpennebulaVirtualMachineRead(d *schema.ResourceData, meta interface
 	err = flattenTemplate(d, &vm.Template)
 	if err != nil {
 		return err
-	}
-	return nil
-}
-
-func flattenTemplate(d *schema.ResourceData, vmTemplate *vm.Template) error {
-
-	var err error
-
-	// VM Group
-	vmgMap := make([]map[string]interface{}, 0, 1)
-	vmgIdStr, _ := vmTemplate.GetStrFromVec("VMGROUP", "VMGROUP_ID")
-	vmgid, _ := strconv.ParseInt(vmgIdStr, 10, 32)
-	vmgRole, _ := vmTemplate.GetStrFromVec("VMGROUP", "ROLE")
-
-	// OS
-	osMap := make([]map[string]interface{}, 0, 1)
-	arch, _ := vmTemplate.GetOS(vmk.Arch)
-	boot, _ := vmTemplate.GetOS(vmk.Boot)
-
-	// Graphics
-	graphMap := make([]map[string]interface{}, 0, 1)
-	listen, _ := vmTemplate.GetIOGraphic(vmk.Listen)
-	port, _ := vmTemplate.GetIOGraphic(vmk.Port)
-	t, _ := vmTemplate.GetIOGraphic(vmk.GraphicType)
-	keymap, _ := vmTemplate.GetIOGraphic(vmk.Keymap)
-
-	// Disks
-	diskList := make([]interface{}, 0, 1)
-
-	// Nics
-	nicList := make([]interface{}, 0, 1)
-
-	// Set OVM Group to resource
-	if vmgIdStr != "" {
-		vmgMap = append(vmgMap, map[string]interface{}{
-			"vmgroup_id": vmgid,
-			"role":       vmgRole,
-		})
-		err = d.Set("vmgroup", vmgMap)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Set OS to resource
-	if arch != "" {
-		osMap = append(osMap, map[string]interface{}{
-			"arch": arch,
-			"boot": boot,
-		})
-		err = d.Set("os", osMap)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Set Graphics to resource
-	if port != "" {
-		graphMap = append(graphMap, map[string]interface{}{
-			"listen": listen,
-			"port":   port,
-			"type":   t,
-			"keymap": keymap,
-		})
-		err = d.Set("graphics", graphMap)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Set Disks to Resource
-	for _, disk := range vmTemplate.GetDisks() {
-		size, _ := disk.GetI(shared.Size)
-		driver, _ := disk.Get(shared.Driver)
-		target, _ := disk.Get(shared.TargetDisk)
-		imageId, _ := disk.GetI(shared.ImageID)
-
-		diskList = append(diskList, map[string]interface{}{
-			"image_id": imageId,
-			"size":     size,
-			"target":   target,
-			"driver":   driver,
-		})
-	}
-
-	if len(diskList) > 0 {
-		err = d.Set("disk", diskList)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Set Nics to resource
-	for i, nic := range vmTemplate.GetNICs() {
-		sg := make([]int, 0)
-		ip, _ := nic.Get(shared.IP)
-		mac, _ := nic.Get(shared.MAC)
-		physicalDevice, _ := nic.GetStr("PHYDEV")
-		network, _ := nic.Get(shared.Network)
-		nicId, _ := nic.ID()
-
-		model, _ := nic.Get(shared.Model)
-		networkId, _ := nic.GetI(shared.NetworkID)
-		securityGroupsArray, _ := nic.Get(shared.SecurityGroups)
-
-		sgString := strings.Split(securityGroupsArray, ",")
-		for _, s := range sgString {
-			sgInt, _ := strconv.ParseInt(s, 10, 32)
-			sg = append(sg, int(sgInt))
-		}
-
-		nicList = append(nicList, map[string]interface{}{
-			"ip":              ip,
-			"mac":             mac,
-			"network_id":      networkId,
-			"physical_device": physicalDevice,
-			"network":         network,
-			"nic_id":          nicId,
-			"model":           model,
-			"security_groups": sg,
-		})
-		if i == 0 {
-			d.Set("ip", ip)
-		}
-	}
-
-	if len(nicList) > 0 {
-		err = d.Set("nic", nicList)
-		if err != nil {
-			return err
-		}
 	}
 	return nil
 }
@@ -816,125 +518,7 @@ func generateVm(d *schema.ResourceData, tplContext *dyn.Vector) (string, error) 
 		}
 	}
 
-	//Generate NIC definition
-	nics := d.Get("nic").([]interface{})
-	log.Printf("Number of NICs: %d", len(nics))
-
-	for i := 0; i < len(nics); i++ {
-		nicconfig := nics[i].(map[string]interface{})
-		nic := tpl.AddNIC()
-
-		for k, v := range nicconfig {
-
-			if k == "network_id" {
-				nic.Add(shared.NetworkID, strconv.Itoa(v.(int)))
-				continue
-			}
-
-			if isEmptyValue(reflect.ValueOf(v)) {
-				continue
-			}
-
-			switch k {
-			case "ip":
-				nic.Add(shared.IP, v.(string))
-			case "mac":
-				nic.Add(shared.MAC, v.(string))
-			case "model":
-				nic.Add(shared.Model, v.(string))
-			case "physical_device":
-				nic.Add("PHYDEV", v.(string))
-			case "security_groups":
-				nicsecgroups := ArrayToString(v.([]interface{}), ",")
-				nic.Add(shared.SecurityGroups, nicsecgroups)
-			}
-		}
-
-	}
-
-	//Generate DISK definition
-	disks := d.Get("disk").([]interface{})
-	log.Printf("Number of disks: %d", len(disks))
-
-	for i := 0; i < len(disks); i++ {
-
-		diskconfig := disks[i].(map[string]interface{})
-		disk := tpl.AddDisk()
-
-		for k, v := range diskconfig {
-
-			if isEmptyValue(reflect.ValueOf(v)) {
-				continue
-			}
-
-			switch k {
-			case "target":
-				disk.Add(shared.TargetDisk, v.(string))
-			case "driver":
-				disk.Add(shared.Driver, v.(string))
-			case "size":
-				disk.Add(shared.Size, strconv.Itoa(v.(int)))
-			case "image_id":
-				disk.Add(shared.ImageID, strconv.Itoa(v.(int)))
-			}
-		}
-	}
-
-	//Generate GRAPHICS definition
-	graphics := d.Get("graphics").(*schema.Set).List()
-	for i := 0; i < len(graphics); i++ {
-		graphicsconfig := graphics[i].(map[string]interface{})
-
-		for k, v := range graphicsconfig {
-
-			if isEmptyValue(reflect.ValueOf(v)) {
-				continue
-			}
-
-			switch k {
-			case "listen":
-				tpl.AddIOGraphic(vmk.Listen, v.(string))
-			case "type":
-				tpl.AddIOGraphic(vmk.GraphicType, v.(string))
-			case "port":
-				tpl.AddIOGraphic(vmk.Port, v.(string))
-			case "keymap":
-				tpl.AddIOGraphic(vmk.Keymap, v.(string))
-			}
-
-		}
-	}
-
-	//Generate OS definition
-	os := d.Get("os").(*schema.Set).List()
-	//vmos := make([]vmOs, len(os))
-	for i := 0; i < len(os); i++ {
-		osconfig := os[i].(map[string]interface{})
-		tpl.AddOS(vmk.Arch, osconfig["arch"].(string))
-		tpl.AddOS(vmk.Boot, osconfig["boot"].(string))
-	}
-
-	//Generate VM Group definition
-	vmgroup := d.Get("vmgroup").(*schema.Set).List()
-	for i := 0; i < len(vmgroup); i++ {
-		vmgconfig := vmgroup[i].(map[string]interface{})
-		vmgroupTpl := tpl.AddVector("VMGROUP")
-		vmgroupTpl.AddPair("VMGROUP_ID", vmgconfig["vmgroup_id"].(int))
-		vmgroupTpl.AddPair("ROLE", vmgconfig["role"].(string))
-	}
-
-	vmcpu, ok := d.GetOk("cpu")
-	if ok {
-		tpl.CPU(vmcpu.(float64))
-	}
-	vmmemory, ok := d.GetOk("memory")
-	if ok {
-		tpl.Memory(vmmemory.(int))
-	}
-	vmvcpu, ok := d.GetOk("vcpu")
-	if ok {
-		tpl.VCPU(vmvcpu.(int))
-	}
+	generateVMTemplate(d, tpl)
 
 	tplStr := tpl.String()
 	log.Printf("[INFO] VM definition: %s", tplStr)